+++ packages/gatsby/src/redux/reducers/config.ts
@@ -1,0 +1,14 @@
+import { IGatsbyConfig, ISetSiteConfig } from "../types"
+
+module.exports = (
+  state: IGatsbyConfig = {},
+  action: ISetSiteConfig
+): IGatsbyConfig => {
+  switch (action.type) {
+    case `SET_SITE_CONFIG`: {
+      return action.payload
+    }
+    default:
+      return state
+  }
+}