+++ www/src/templates/tags.js
@@ -1,6 +1,8 @@
 /** @jsx jsx */
 import { jsx } from "theme-ui"
 import React from "react"
+import styled from "@emotion/styled"
+import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
 import { graphql } from "gatsby"
 import { TiTags as TagsIcon, TiArrowRight } from "react-icons/ti"
 import Main from "../components/main-content"
@@ -24,6 +26,16 @@ const preferSpacedTag = tags => {
   return tags[0]
 }
 
+const ButtonWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+
+  ${mediaQueries.xs} {
+    flex-direction: row;
+  }
+`
+
 const Tags = ({ pageContext, data }) => {
   const { tags } = pageContext
   const { nodes, totalCount } = data.allMdx
@@ -33,13 +45,42 @@ const Tags = ({ pageContext, data }) => {
   const doc = TAGS_AND_DOCS.get(tags[0])
 
   return (
-    <Main>
-      <Container>
-        <PageMetadata
-          title={`${preferSpacedTag(tags)} Tag`}
-          description={`Case studies, tutorials, and other posts about Gatsby related to ${preferSpacedTag(
-            tags
-          )}`}
+
+    <Container>
+      <PageMetadata
+        title={`${preferSpacedTag(tags)} Tag`}
+        description={`Case studies, tutorials, and other posts about Gatsby related to ${preferSpacedTag(
+          tags
+        )}`}
+      />
+      <h1>{tagHeader}</h1>
+      <ButtonWrapper>
+        <Button
+          variant="small"
+          key="blog-post-view-all-tags-button"
+          to="/blog/tags"
+        >
+          View all tags <TagsIcon />
+        </Button>
+        {doc ? (
+          <React.Fragment>
+            <span css={{ margin: 5 }} />
+            <Button
+              variant="small"
+              secondary
+              key={`view-tag-docs-button`}
+              to={doc}
+            >
+              Read the documentation <TiArrowRight />
+            </Button>
+          </React.Fragment>
+        ) : null}
+      </ButtonWrapper>
+      {nodes.map(node => (
+        <BlogPostPreviewItem
+          post={node}
+          key={node.fields.slug}
+          sx={{ my: 9 }}
         />
         <h1>{tagHeader}</h1>
         <Button
@@ -84,8 +125,7 @@ export const pageQuery = graphql`
       sort: { fields: [frontmatter___date, fields___slug], order: DESC }
       filter: {
         frontmatter: { tags: { in: $tags } }
-        fileAbsolutePath: { regex: "/docs.blog/" }
-        fields: { released: { eq: true } }
+        fields: { section: { eq: "blog" }, released: { eq: true } }
       }
     ) {
       totalCount