+++ packages/gatsby-plugin-preact/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@3.3.0...gatsby-plugin-preact@3.3.1) (2020-05-05)
+
+**Note:** Version bump only for package gatsby-plugin-preact
+
 # [3.3.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@3.2.4...gatsby-plugin-preact@3.3.0) (2020-04-27)
 
 **Note:** Version bump only for package gatsby-plugin-preact