+++ packages/gatsby-plugin-typography/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.5.0...gatsby-plugin-typography@2.5.1) (2020-05-05)
+
+**Note:** Version bump only for package gatsby-plugin-typography
+
 # [2.5.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.4.4...gatsby-plugin-typography@2.5.0) (2020-04-27)
 
 **Note:** Version bump only for package gatsby-plugin-typography